feat: check functions in order of priority
This allows the execution order of the overloaded functions to be
controlled, in cases where such control is necessary.

This commit also refactors the core dispatching functionality into its
own function: `signature_dispatch.dispatch()`.  I also decided to expose
this function as part of the public API, because it seems generally
useful.